Updated Test
Updated the user.delete tests

+++ routes/users.js
@@ -69,12 +69,13 @@ module.exports = () => {
   router.delete('/:userId', authorizeSession, async (req, res, next) => {
     try {
       const userId = req.params.userId;
+      if (!userId || userId === '') {
+        throw HttpError(400, 'Bad Parameters');
+      }
       let user = await User.findOne({ userId: userId });
+
       if (isEmpty(user)) {
         throw new HttpError.NotFound();
-      }
-      if (!userId) {
-        throw HttpError(400, 'Bad Parameters');
       }
 
       user = await User.deleteUser(userId);
